requirements.txt
google-cloud-bigquery>=3.10.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
db-dtypes>=1.2.0
